refactor: mutualize TCP and Default ApiReactors https://gravitee.atlassian.net/browse/APIM-3228
